+++ meerschaum/api/dash/pages/__init__.py
@@ -11,3 +11,4 @@ import meerschaum.api.dash.pages.dashboard
 import meerschaum.api.dash.pages.plugins
 import meerschaum.api.dash.pages.register
 import meerschaum.api.dash.pages.pipes
+import meerschaum.api.dash.pages.job